rename: 자판기 컴포넌트 경로 변경
