+++ fe/src/Components/MoneyScreenBoard/ChargeForm.jsx
@@ -1,6 +1,6 @@
 import { SetWalletMoneyContext, WalletMoneyContext } from 'App';
 import { INIT_ALERT_MESSAGE } from 'Helper/constant';
-import { InvestmentContext, SetAlertMessage, SetInvestmentContext } from 'Pages/VendingMachine';
+import { InvestmentContext, SetAlertMessage, SetInvestmentContext } from 'Pages/VendingMachine/VendingMachine';
 import { useContext, useState } from 'react';
 import { CashInput, Button, ChargeForm } from './ChargeForm.styled';
