+++ fe/src/Pages/VendingMachine/VendingMachine.jsx
